# Conflicts:
#	.yarn/install-state.gz
#	common/model/db/tables/dataNodeDef/table.js
#	core/objectUtils.js
#	core/record/_record/recordNodesUpdater.js
#	core/survey/surveyFile.js
#	package.json
#	server/modules/collectImport/service/collectImport/dataImportJobs/collectAttributeValueExtractor.js
#	server/modules/dataImport/service/DataImportJob/EntitiesDeleteJob.js
#	server/modules/mobile/service/arenaMobileDataImport/jobs/recordsImportJob.js
#	server/modules/record/api/recordApi.js
#	server/modules/record/manager/_recordManager/nodeCreationManager.js
#	server/modules/record/manager/_recordManager/nodeUpdateManager.js
#	server/modules/record/manager/recordManager.js
#	server/modules/record/service/recordService.js
#	server/modules/record/service/recordsCloneJob.js
#	server/modules/record/service/update/thread/recordsUpdateThread.js
#	server/modules/survey/service/recordCheckJob.js
#	server/modules/surveyRdb/repository/dataTableInsertRepository.js
#	server/modules/surveyRdb/repository/dataTableUpdateRepository.js
#	server/modules/surveyRdb/repository/dataView/create.js
#	server/modules/surveyRdb/service/surveyRdbCreationJob/recordsUniquenessValidationJob.js
#	test/e2e/tests/_surveyExport/records.js
#	webapp/components/survey/NodeDefsSelector/NodeDefTreeSelect/useBuildEntityTreeData.js
#	webapp/components/survey/Record/RecordNoHeader.js
#	webapp/components/survey/SurveyForm/nodeDefs/components/types/nodeDefEntityFormNodeSelect.js
#	webapp/components/survey/SurveyForm/nodeDefs/components/types/nodeDefEntityTableRows.js
#	webapp/store/ui/record/hooks/index.js
#	webapp/store/ui/surveyForm/state.js
#	webapp/views/App/views/Data/MapView/MapView.js
#	yarn.lock

+++ server/modules/auth/api/authApi.js
@@ -8,6 +8,7 @@ import * as Survey from '@core/survey/survey'
 import * as User from '@core/user/user'
 import * as Authorizer from '@core/auth/authorizer'
 import * as ProcessUtils from '@core/processUtils'
+import SystemError from '@core/systemError'
 
 import * as Log from '@server/log/log'
 import * as SurveyService from '../../survey/service/surveyService'
@@ -16,7 +17,7 @@ import * as RecordService from '../../record/service/recordService'
 
 const Logger = Log.getLogger('AuthAPI')
 
-const sendResponse = (res, user, survey = null) => res.json({ user, survey })
+const sendResponse = (res, user, survey = null, error = null) => res.json({ user, survey, error })
 
 const sendUserSurvey = async (res, user, surveyId) => {
   try {
@@ -28,8 +29,15 @@ const sendUserSurvey = async (res, user, surveyId) => {
     sendResponse(res, user, survey)
   } catch (error) {
     Logger.error(`error loading survey with id ${surveyId}: ${error.toString()}`)
-    // Survey not found with user pref
-    // removing user pref
+
+    if (error instanceof SystemError && error.key === 'survey.dataMigrationInProgress') {
+      // Survey data migration still in progress: keep the user's current survey preference untouched,
+      // it's not a case of the survey being missing or inaccessible, just temporarily unavailable.
+      sendResponse(res, user, null, { key: error.key, params: error.params })
+      return
+    }
+
+    // Survey not found (or otherwise inaccessible) with user pref: removing user pref
     const _user = User.deletePrefSurvey(surveyId)(user)
     sendResponse(res, await UserService.updateUserPrefsAndFetchGroups(_user))
   }
@@ -49,12 +57,14 @@ export const init = (app) => {
   app.get('/auth/user', async (req, res, next) => {
     try {
       const user = Request.getUser(req)
+      const surveysCount = await UserService.fetchUserSurveysCount(User.getUuid(user))
+      const userWithCounts = { ...user, ...surveysCount }
       const surveyId = User.getPrefSurveyCurrent(user)
 
       if (surveyId) {
-        await sendUserSurvey(res, user, surveyId)
+        await sendUserSurvey(res, userWithCounts, surveyId)
       } else {
-        sendResponse(res, user)
+        sendResponse(res, userWithCounts)
       }
     } catch (error) {
       next(error)
@@ -65,7 +75,7 @@ export const init = (app) => {
     try {
       // Before logout checkOut record if there's an opened thread
       const socketId = Request.getSocketId(req)
-      RecordService.dissocSocketFromUpdateThread(socketId)
+      await RecordService.dissocSocketFromUpdateThread(socketId)
 
       res.clearCookie('refreshToken', {
         httpOnly: true,
